chore: reorganizo mappers moviéndolos de utils a data/mapper
